requirements.txt
python-dotenv
llama-index
html2text
black
pinecone-client
streamlit